Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Yes, because everyone want a slightly slower running phone just to cater to building on a bastardized version of BSD. ;)

But seriously, like I was about to say earlier before Steve already did it this was added at least weeks ago so must be something else. If this was really the culprit though, couldn't it be ifdef:ed, a built flag or something like that? I would assume something like 90% of people building it themselves use some form of Linux and not OS X. Just saying.

Patch-set: 1