Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

If this is to be included, it will probably be better to fork it, have it replace libjpeg in the manifest and have it provide libjpeg. This will lose all development history for libjpeg-turbo and will make future merges a pain.

Patch-set: 3